add bigList and fix checking system
The data structure bigList has been added, which is a list of lists of
Data objects. The function printBigList has also been added, which
prints out this data structure. The system for checking if two data
objects can be combined has also been updated, it now only checks if two
objects whose difference of the count of 1s in binary is only one.

Example:
10010 and 00000 can not be combined because 10010 has two "1"s and 00000
has zero ones, and this difference is greater than one so these can
not be combined.